// src/modules/comments/dto/comment.view-dto.ts
import { LikeStatus } from '../domain/comment.entity';

export type CommentViewDto = {
  id: string;
  content: string;
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  createdAt: string;
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: LikeStatus;
  };
};